API Connection Token - Permissions update.
Users can now choose the permissions their API Connection Token has. They can grant to publish, read, or delete private notes.

+++ api/token/delete-private-note.php
@@ -1,6 +1,6 @@
 <?php
 
-//Route: onlinenotes.vice/api/token/insert-private-note.php
+//Route: onlinenotes.vice/api/token/delete-private-note.php
 //JSON template:
 /*
 {
@@ -20,20 +20,37 @@ require "../../app/database/delete.php";
 require "../../app/database/read.php";
 
 $data = json_decode(file_get_contents("php://input")); //Gets data from request JSON.
-$noteid = $data->noteid;
-$token = $data->token;
 
-if (token_given($token)) {
-    $user_id = tokenValid_returnId($token);
-    if ($user_id != null) {
-        delete_private_note_both_ids($noteid, $user_id);
+if (isset($data->noteid)) {
+    $noteid = $data->noteid;
+    if (isset($data->token)) {
+        $token = $data->token;
+        if (token_given($token)) {
+            $user_id = tokenValid_returnUserId($token);
+            if ($user_id != null) {
+                if (tokenHasDeletePermission($token)) {
+                    delete_private_note_both_ids($noteid, $user_id);
+                    echo json_encode(
+                        array("status" => "success", "description" => "query executed"),
+                        JSON_PRETTY_PRINT
+                    );
+                }
+            }
+        }
+    } else {
         echo json_encode(
-            array("status" => "success", "description" => "query executed"),
+            array("status" => "failed", "description" => "token not given"),
             JSON_PRETTY_PRINT
         );
     }
+} else {
+    echo json_encode(
+        array("status" => "failed", "description" => "note id not given"),
+        JSON_PRETTY_PRINT
+    );
 }
 
+// Checks if token content was given.
 function token_given($token)
 {
     if ($token != "") {
@@ -49,9 +66,9 @@ function token_given($token)
 // Checks if the token is valid.
 // If valid, returns user's ID.
 // If not valid, prints error code.
-function tokenValid_returnId($token)
+function tokenValid_returnUserId($token)
 {
-    $return_obj = array("status" => "false", "ID" => "token not valid");
+    $return_obj = array("status" => "failed", "description" => "token not valid");
     $obj = api_connection_token_brind_id_if_exists($token);
 
     if ($obj == null) {
@@ -61,5 +78,18 @@ function tokenValid_returnId($token)
         return null;
     } else {
         return $obj["user_id"];
+    }
+}
+
+// Returns true if so. Prints error if no read permission.
+function tokenHasDeletePermission($token)
+{
+    $tokenPermissions = apiConnectionToken_bringPermissionDetails($token);
+    if ($tokenPermissions["DeletePermission"] == 1) return true;
+    else {
+        echo json_encode(
+            array("status" => "failed", "description" => "delete permission disallowed")
+        );
+        return null;
     }
 }
